Add chartMixin for cleaning the code
