+++ client/pages/bookmark.vue
@@ -73,49 +73,20 @@
 
 <script>
 import axios from 'axios';
+import chartMixin from '@/mixins/chart-mixin';
 
 const protocol = process.env.PROTOCOL;
 const origin = process.env.ORIGIN;
 const bookmarkURL = `${protocol}://${origin}/api/bookmark/`;
 
-const HTTP_201_CREATED = 201;
-const HTTP_204_NO_CONTENT = 204;
-
 export default {
+    mixins: [chartMixin],
     beforeRouteUpdate (to, from, next) {
         // クエリなしに対応
         this.query.ordering = to.query.ordering || '';
         window.scrollTo(0, 0);
         this.fetchBookmarkAPIData();
         next();
-    },
-    data () {
-        return {
-            // Main Card
-            items: [],
-            bookMarkIDs: [],
-            isVisible: false,
-
-            chartGridCol: 12,
-            chartHight: 300,
-            sortItems: [
-                { text: '新着順', value: '' },
-                { text: 'GPA (降順)', value: '-gpa' },
-                { text: 'GPA (昇順)', value: 'gpa' },
-                { text: '単位取得', value: 'failure' },
-                { text: '落単', value: '-failure' },
-                { text: 'A帯 (降順)', value: '-a_band' },
-                { text: 'A帯 (昇順)', value: 'a_band' },
-                { text: 'F (降順)', value: '-f' }
-            ],
-            gridItems: [
-                { text: '１列', value: 12 },
-                { text: '２列', value: 6 }
-            ],
-            query: {
-                ordering: ''
-            }
-        };
     },
     head () {
         return {
@@ -135,52 +106,6 @@ export default {
         this.chartGridCol = window.innerWidth <= 600 ? 12 : 6;
     },
     methods: {
-        sort () {
-            const fullURL = this.joinQuery(this.$route.path);
-            this.$router.push(fullURL);
-        },
-
-        joinQuery (url) {
-            let queryURL = '';
-            Object.keys(this.query).forEach((key) => { queryURL += '&' + key + '=' + this.query[key]; });
-            if (!url.includes('?')) {
-                queryURL = queryURL.replace('&', '?'); // 先頭の&を?に置換
-            }
-            return url + queryURL;
-        },
-
-        async postBookMark (index) {
-            // Objectをコピーする必要がある。
-            const item = Object.assign({}, this.items[index]);
-            const bookMarkID = item.id;
-
-            if (item.isBookMark) {
-                // ブックマーク解除
-                const res = await axios.delete(`${bookmarkURL}${bookMarkID}/`, {
-                    withCredentials: true
-                });
-                if (res.status === HTTP_204_NO_CONTENT) {
-                    this.bookMarkIDs = this.bookMarkIDs.filter(id => id !== bookMarkID);
-                    item.isBookMark = !item.isBookMark;
-                    this.$set(this.items, index, item);
-                }
-            } else {
-                // 登録
-                const res = await axios.post(
-                    bookmarkURL,
-                    { bookMarkID },
-                    {
-                        withCredentials: true
-                    }
-                );
-
-                if (res.status === HTTP_201_CREATED) {
-                    this.bookMarkIDs = res.data.bookMarkIDs;
-                    item.isBookMark = !item.isBookMark;
-                    this.$set(this.items, index, item);
-                }
-            }
-        },
 
         async fetchBookmarkAPIData () {
             this.isVisible = false;
