+++ client/pages/search.vue
@@ -147,16 +147,15 @@
 <script>
 import axios from 'axios';
 import MobileDetect from 'mobile-detect';
+import chartMixin from '@/mixins/chart-mixin';
 
 const protocol = process.env.PROTOCOL;
 const origin = process.env.ORIGIN;
 const gradeURL = `${protocol}://${origin}/api/gradeinfo/`;
 const bookmarkURL = `${protocol}://${origin}/api/bookmark/`;
 
-const HTTP_201_CREATED = 201;
-const HTTP_204_NO_CONTENT = 204;
-
 export default {
+    mixins: [chartMixin],
     beforeRouteUpdate (to, from, next) {
     // クエリなしに対応
         this.query.page = to.query.page || 1;
@@ -168,10 +167,6 @@ export default {
     },
     data () {
         return {
-            // MainCard
-            items: [],
-            bookMarkIDs: [],
-            isVisible: false,
             // Pagination
             currentPage: 1,
             totalVisible: 0,
@@ -179,28 +174,7 @@ export default {
             count: null,
 
             search: '',
-            searchResultText: null,
-            chartGridCol: 12,
-
-            query: {
-                search: '',
-                ordering: '',
-                page: 1
-            },
-            sortItems: [
-                { text: '新着順', value: '' },
-                { text: 'GPA (降順)', value: '-gpa' },
-                { text: 'GPA (昇順)', value: 'gpa' },
-                { text: '単位取得', value: 'failure' },
-                { text: '落単', value: '-failure' },
-                { text: 'A帯 (降順)', value: '-a_band' },
-                { text: 'A帯 (昇順)', value: 'a_band' },
-                { text: 'F (降順)', value: '-f' }
-            ],
-            gridItems: [
-                { text: '１列', value: 12 },
-                { text: '２列', value: 6 }
-            ]
+            searchResultText: null
         };
     },
     async fetch () {
@@ -215,12 +189,12 @@ export default {
 
         if (process.server) {
             const headers = this.$nuxt.context.req.headers;
-            await this.fetchBookmarkAPIdata(headers);
+            await this.fetchBookmarkAPIData(headers);
             const md = new MobileDetect(headers['user-agent']);
             this.chartGridCol = md.mobile() ? 12 : 6;
             this.totalVisible = md.mobile() ? 5 : 10;
         } else {
-            await this.fetchBookmarkAPIdata();
+            await this.fetchBookmarkAPIData();
         }
     },
     head () {
@@ -244,13 +218,6 @@ export default {
         filterSearch () {
             // vuetifyのclearはnullが挿入される
             this.query.search = this.search || '';
-            this.query.page = 1;
-
-            const fullURL = this.joinQuery(this.$route.path);
-            this.$router.push(fullURL);
-        },
-
-        sort () {
             this.query.page = 1;
 
             const fullURL = this.joinQuery(this.$route.path);
@@ -294,7 +261,7 @@ export default {
         },
 
         // ページに訪れた時のみに使用
-        async fetchBookmarkAPIdata (headers) {
+        async fetchBookmarkAPIData (headers) {
             const options = {
                 withCredentials: true
             };
@@ -304,50 +271,6 @@ export default {
             this.items.forEach((item) => {
                 this.$set(item, 'isBookMark', this.bookMarkIDs.includes(item.id));
             });
-        },
-
-        async postBookMark (index) {
-            // Objectをコピーする必要がある。
-            const item = Object.assign({}, this.items[index]);
-            const bookMarkID = item.id;
-
-            if (item.isBookMark) {
-                // ブックマーク解除
-                const res = await axios.delete(`${bookmarkURL}${bookMarkID}/`, {
-                    withCredentials: true
-                });
-                if (res.status === HTTP_204_NO_CONTENT) {
-                    this.bookMarkIDs = this.bookMarkIDs.filter(id => id !== bookMarkID);
-                    item.isBookMark = !item.isBookMark;
-                    this.$set(this.items, index, item);
-                }
-            } else {
-                // 登録
-                const res = await axios.post(
-                    bookmarkURL,
-                    { bookMarkID },
-                    {
-                        withCredentials: true
-                    }
-                );
-
-                if (res.status === HTTP_201_CREATED) {
-                    this.bookMarkIDs = res.data.bookMarkIDs;
-                    item.isBookMark = !item.isBookMark;
-                    this.$set(this.items, index, item);
-                }
-            }
-        },
-
-        joinQuery (url) {
-            let queryURL = '';
-            Object.keys(this.query).forEach(
-                key => (queryURL += '&' + key + '=' + encodeURIComponent(this.query[key]))
-            );
-            if (!url.includes('?')) {
-                queryURL = queryURL.replace('&', '?'); // 先頭の&を?に置換
-            }
-            return url + queryURL;
         }
     }
 };
